Move logo key to public method
Unfortunately the letter branding flows will need to be able to retrieve
the full S3 object key (filepath) for a given logo file, so we need to
make this public.

`letter_branding.filename` stores the base name of a file excluding the
extension. We store all letter logos in an S3 bucket under
`letters/static/images/letter-template/blah.svg`, but would only store
`blah` in the DB. The letter template previews hard-code the rest of the
path in the HTML template.

When we're editing letter brands in admin, we need to be able to get the
full path of the S3 object and we can do that by getting the logo key
from the logo client. In an ideal world we would probably store the full
logo key (absolute path) in the DB, but the migration now isn't worth
the effort as there's a lot of other places that tie into this.