Split out the org branding code into its own file
We had all of the code to do with organisations in
"app/main/views/organisations.py", but this file was getting long and
covering lots of different areas. This change creates a new
"organisations" directory containting the existing file now split into
two separate files - one for branding and one for everything else.