Split out user facing email branding code into its own file
All the email branding code was in
"app/main/views/service_settings/index.py" - both the platform admin
only views and the views that allow users to change their branding
themselves.

This keeps the platform admin views in that file - there are not many and
they relate to both email and letter branding.

The user-facing email branding code has been moved to
- app/main/views/service_settings/email_branding.py (new file)
- tests/app/main/views/service_settings/test_email_branding_requests.py
  (already existed but did not contain all the tests)