Allow services to enable email auth
Service managers can request to allow their users to login via a magic
email link instead of a text message as a form of 2FA. At the moment
this goes through Zendesk, requiring support action from us, but we
believe this should be able to be self-served.

This changes the `set-auth-type` change to a form which allows managers
to decide between text message and magic link auth.

Enabling email auth does not change users' sign in to actually use email
links - so service managers will still need to go to the team members
page and update permissions individually from there. A future patch will
provide an immediate page where the manager can update permissions in
bulk.

Disabling email auth _does_ automatically change any users of the service
who sign-in with emails to sign in with text messages instead. Anyone
signing in with webauthn is unaffected.