Add sign out link to profile page
We’re going to remove this link from the header. So that users can still
sign out, let’s add it to the profile page instead.

We think that most users don’t use the sign out[1] link so it doesn’t
need to be so prominent.

We can better use the space for some of our guidance in future PRs.

1. ~13,000 Kibana hits to `/sign-in*` versus ~500 to `/sign-out`