Remove sign out links
We think that most users don’t use the sign out[1] link so it doesn’t
need to be so prominent.

We’ve already added it to the profile page, so there is still a way for
users who manually want to sign out to do so.

We can better use the space for some of our guidance in future PRs.

1. ~13,000 Kibana hits to `/sign-in*` versus ~500 to `/sign-out`