Add a central 'find anything' page
Uses the new `find-by-uuid` API endpoint for platform admins to provide
a single central search form that can take any UUID and send a platform
admin to the best page we have to view that record.

As part of this change, we pull the two other 'platform admin lookup'
pages onto a single page so that we don't have 3 pages for looking up
records. This page is fairly unoptimised and shows 3 search boxes. The
existing ones (service and user) redirect to their existing pages, while
the find-by-uuid takes the user to the actual record page where
appropriate.