Allow non-emergency 3-digit contact numbers
NHS COVID Pass want their 'send a file' contact phone number to be 119,
but we currently require numbers to be 5 digits or more. This is clearly
not the right restriction to have, as 119 is legitimately the contact
point for covid enquiries.

Let's allow 3-digit (and technically 4-digit numbers) numbers as points
of contact, but specifically exclude emergency numbers that don't seem
reasonable to use.