create new redirects blueprint to manage old redirects
there's a bit of boilerplate and cruft around managing old redirects
that is largely unnecessary - we shouldn't care about navigation
exclusions for old endpoints that just return a 301 to a new page.
We've also ended up with some redirect chains that are a bit unnecessary
- the really old endpoints can just link straight to the current
  endpoints, rather than to intermediate steps that are also old and
outdated.

Note that we're kinda abusing werkzeug routing's defaults dict. Normally
it's designed for a case where you have routes like:

```
@app.route("/things/", default={"id": 1})
@app.route("/things/<id>")
def get_thing(id):
   ...
```

with `redirect_defaults` enabled, this says that the "/things/" endpoint
with no ID is the canonical URL if you enter id=1 - so navigation to
/things/1 will return a 308 permanent redirect. That's not quite what we
want here, where we want to specify "here's a route, when you hit this
particular URL, call the route function with a specific arbitrary
argument". By turning off the redirect_defaults flag, it populates the
route argument nicely, allowing us to know which URL the user hit
without having to define different view functions for every redirect.