let accessibility_statement test handle non-functional changes to the file
previously it woudl error if you didn't update the last updated at
timestamp, even if you hadn't changed the doc itself and hadn't actually
assessed the statement.

Now there's a timestamp in a jinja comment for when the last
non-functional change was. You must update either this or the last
updated at timestamp. Still not perfect but a bit less brittle.

Also call git diff with `--unified=0` to not return any context, so we
aren't getting false positives on context lines of the git diff