Allow new template permissions to be toggled by platform admin
We've added two new service permissions to be feature flags:
- `extra_email_formatting`
- `extra_letter_formatting`

This provides a way for a platform admin user to toggle these
permissions from the service settings page. The option to toggle
`extra_email_formatting` will only be visible if the service has the
`email` permission and the option to toggle `extra_letter_formatting`
will only be visible if the service has the `letter` permission.