Remove international letters row from platform admin settings
International letters used to be an invite only feature, but it is now
the default for all services and something they can turn on and off
themselves from the settings page.

Since there is now a row to toggle the permission in the "Text message
settings" section, we no longer also need a row which does the same thing
in 'Platform admin settings' section.