Move a final redirect option in branding options view out of
conditional statement.

Reasoning:

It's sort of more explicit, but also a little bit semantically different. If the final option is inside an else block, then the fallback state for form.validate_on_submit becomes "render this template". If it's not inside the else block, then we are terminating the if form.validate_on_submit with a redirect. It's a bit more defensive against future changes. With the code as it stands, it doesn't make a difference because we know that all of the branches of the if explicitly return. But if one of them didn't, we'd fall out of the validated form flow altogether and end up back in the basic "get page template render".

Co-authored-by: Sam Williams <sam.williams@digital.cabinet-office.gov.uk>