Fix vertical alignment of performance page
This is complex because it involved aligning
content in the normal flow of the document with
that in data tables.

The changes made are:
- make the 'total amounts' section use `display:
  table` for layout
- change how the gutters are applied for both
  content items and table cells and set to the
  same size for all
- reduce the width of the first column at various
  screen sizes to give more room to the other
  cells (or content items)

Each of these requires more explanation.

1. Make the 'total amounts' section use `display:
table` for layout.

I tried CSS grid and flexbox. Both have
differences:
- flexbox with a set width on the first item will
  reduce it at smaller sizes to allow all items to
  fit, unlike tables
- flexbox with a min-width on the first item will
  increase the width of the parent container when
  it runs out of horizontal space, unlike tables
- grid with a set width on the first item will
  increase the width of the parent container when
  it runs out of horizontal space, unlike tables

Both grid and flexbox also seem to use a slightly
different rounding for percentages in some
browsers to tables, leading to a few pixel
difference with the same width in tables.

Setting `display: table` on the 'total amounts'
makes them display closest to the table below but
it does have its own differences:
- it applies width separate to padding, whereas
  table headers include the padding in whatever
  width is set, though this, again, varies across
  browsers
- items without width set shrink to their minimum
  content size, taking space away from those with
  width to allow this whereas table headers
  without width set don't disturb those with it,
  they just take a percentage of the width left
  over

2. change how the gutters are applied for both
content items and table cells and set to the same
size for all

Because the 'total amounts' items doesn't include
padding in its width but table headers do, this
moves the padding off the first item and table
cell to guarantee they line up.

3. reduce the width of the first column on various
screen sizes to give more room to the other cells
(or content items)

The combination of font-size and space available
at certain screen sizes makes things break out of
alignment. This is because the items and table
headers without a width set behave differently
when the space left for them is less than their
normal content width.

To help with this, this makes the first item /
table header different widths depending on screen
size. This is targeted to the performance page
only so works because we know the content those
items / table headers will contain.