Visually hint which pages are on the same sheet
Because our letters are double sided, pages 1 and 2; 3 and 4; etc. will
appear on the same sheets of paper.

This commit adds some styling to remove the space between these adjacent
pages. This is a visual cue often used by desktop publishing programs to
show pages which are on the same spread.

Hopefully this will help a little bit once we start supporting
attachments in letters, where it might not be obvious that an attachment
could be printed on the reverse side of a page of the letter.

The `:nth-of-type` CSS selector has good support in all browsers,
including IE11: https://caniuse.com/?search=nth-of-type

I used negative margins so as not to disrupt the spacing after an odd
page of the letter, where it also appears as the last page of the
letter.