Say ‘Untitled’ instead of ‘New’ letter template
When you add a letter template we give it a default name of ‘New letter
template’.

Then when you go to your templates pages you see:
- a link saying ‘New letter template’
- a button saying ‘New template’

In research we’ve seen people confused about how to add another
template, and how to get back to the one they’ve previously been working
on. We think this is because we’re using the same word (new) to refer to
different things.

‘Untitled’ is the word used by Microsoft Windows, Google Docs, Mac OS
and probably others to mean ‘thing that exists but hasn’t been named
yet’.