Add `user_id` to log messages
notifications-utils now checks `g.user_id` and, if present, annotates
log messages with it.

This will make it easier for us to find requests made by specific users
on our logging platform, which will aid debugging and investigation of
reported issues.

We also start to pass through X-Notify-User-Id to all requests to the
API so that we can track which user is actually performing specific
calls there.