Store whole address in recipient field for one off letters
We introduced sending one-off, templates letters from the admin app in
https://github.com/alphagov/notifications-admin/pull/2434/files

Because all notifications require something in the `to` field, we chose
the first line of the address:
https://github.com/alphagov/notifications-admin/commit/3a62946ecd2ab2982d479481f96bf1df487442f8#diff-7365e34b1112e69a28e0c0ff051358e5bb02b41796dc9fadf259f7cff070cc0cR909

At the time this is how other templates letters (sent via API or
spreadsheet upload) worked.

Since then we have changed to instead store the whole address, including
line breaks in the `to` field. Additionally the API populates, the
`normalised_to` field with a lowercase, no whitespace translation which
which we use to search by recipient. This happened in
https://github.com/alphagov/notifications-api/pull/2814/files

However when we made this change we missed out on applying it to one-off
templated letters, which are a bit different because the admin app
passes the address to the API, rather than the API getting it from a
client or from the spreadsheet.

This commit updates the admin app to pass in the whole address, meaning
that one-off templated letters will be searchable like any other.

The API will get this data here:
https://github.com/alphagov/notifications-api/blob/0a407733173b35ebc6a76fcf0870c57fefafa686/app/service/send_notification.py#L95

And create the normalised equivalent here:
https://github.com/alphagov/notifications-api/blob/0a407733173b35ebc6a76fcf0870c57fefafa686/app/notifications/process_notifications.py#L146