Update content following content review.
- don't dynamically tell how long attachment can be - because we cannot
be sure they’ve added all their content to the template. So any kind
of realtime page count could be very misleading.

- remove content about attached pages starting on the next available
page - we want to prove that there’s a need to include it first.

- remove info about required size and layout of the attachment, as
this info is covered in the letter specification, and we do not tell users
about maximum file size at the start of any of our other upload journeys

Co-authored-by: karlchillmaid <karl.chillmaid@digital.cabinet-office.gov.uk>