Show last logged in for each user
When choosing who to contact I reckon it’s useful to know which users
are active (in other words which are likely to respond).

A proxy we can use for this is when they last logged in.

If a user hasn’t logged in for a long time it’s likely because they have
left the service or organisation.