Refactor how polling is stopped
At the moment, the entire module breaks if a single response comes back
with a 500. This adds a stop-tracker per live-updating region, allowing
a single component to break without corrupting the entire module.

Also update the new JSON blueprint to return 401 for unauthorised
access rather than a sign-in page.

For JSON endpoints, we don't want to redirect to the sign-in page if the
session has expired. Instead we should return a plain 401 response,
which javascript can then detect to trigger a page refresh (which will
take the user to the sign-in page).