Handle form errors on submit.
When a user uploads a file that's not a PDF or fails to upload
a file, form validation kicks in. This commit makes sure
we handle this and playback the right information to the user.

Most users actually won't get to that point, as their
browser will not let them upload files outside the spec.