Fix updateContent JS and tests
This makes changes to the tests for the
updateContent JS, to allow us to properly test
scenarios where responses to the requests it makes
complete with either a 401 status code or the
'stop' flag set in the returned data.

This also includes the following changes that came
out of this work:
- fixes for bugs found in the updateContent JS
- changes to how we stop polling the server in
  callbacks
- adding code to tidy up some state left by tests
  that result in requests fired but not completed

Changes to the updateContent JS

We currently only test that the first request
fires when it should and that it is correct.
We do test the results of a valid response but in
a hacky way where callbacks actually fire at the
same time they're attached.
This means:
- we can't really test responses that change
  polling or have side effects
- tests for sequences of requests (like that
  testing the dynamic interval) either won't work
  or are currently working in a way that produces
  false-positives

These changes try to tackle this by:
- adding `serverResponse` to give some control
  over requests and make how they work with
  callbacks more like they do in web pages
- adding code to tests to make certain events
  happen in the same sequence they do in web
  pages, so the behaviour of the async parts of
  updateContent can be tested more accurately

Fix for a bug found in the updateContent JS

When the `poll` function runs for the first time,
the `interval` variable is still set to `0` so it
then calls itself as many times as can be executed
before the first request completes. Setting it to
the `defaultInterval` value when the module starts
means this cannot happen.

Changes to how we stop polling the server in
callbacks

We were changing a flag on the `stopPolling`
object to shut a guard around the call to
`setTimeout` that continues polling.

This works but the code in `poll` before the guard
still executes, causing an additional request to
fire. We instead clear any pending timeout so none
of `poll` runs.

Adding code to tidy up some state left by tests
that result in requests fired but not completed

`poll` has a guard that prevents it firing
requests after the first one, by checking the length
of the renderers queue. That queue is only
emptied, allowing the guard to be bypassed, when
requests complete and their callback called.
Because of this, any requests fired by test code
left incomplete can block future tests. We protect
against this by calling the callback for a failed
request, to tidy up the state.