Catch expired session POST CSRF error
Some of our .json endpoints get POSTed to, which triggers CSRF checks.
If a user's session has expireed, they have no CSRF token in their
session, triggering the CSRFError handler `handle_csrf`. This hands off
to `application.login_manager.unauthorized()`, which - for JSON
endpoints - will abort with HTTPStatus.UNAUTHORIZED. The internals of
this raise a (Werkzeug) error, expected to be picked up by an
errorhandler. However, we're already inside an errohandler, so it
instead causes a 500.

Let's catch the werkzeug error that can be thrown by the login manager
and manually redirect to the no_permissions (401) handler.