Relabel ‘SMS limit’ to ‘Text message limit’
We try to call things ‘text messages’ rather than ‘SMS’ because it is
more widely understood and avoids an initialism. This includes the page
linked to from the settings page, which is titled ‘Daily text message
limit’.

This commit updates the label on the service settings page to be
consistent.