Fix breadcrumb separator alignment (service nav)
Not sure when this went out of alignment but it was appearing too high
and too far to the right. It should be centered within the cap height of
the text to visually look correct.

The service name was also pushed a few pixels too far to the right (not
lining up with the grid columns) when the organisation name overflowed.

This commit shifts the pixels around so everything looks right again
(similar to how it was first designed in https://github.com/alphagov/notifications-admin/pull/3024)