Display 'last updated at' for org dashboard
We are going to stop on-demand calculation of up-to-the-second billing
data when users visit the org dashboard[^1]. Given this, we need to show
an indication of when the data was last generated so that users know if
it's a bit out of date.

We'll add a small "Updated at" line to the dashboard.

[^1]: https://github.com/alphagov/notifications-api/pull/3775