Add a page for renaming letter templates
A lot of the actions for letter templates (editing the content, adding
branding, editing the contact block, etc) are on separate pages.

Renaming the template is still on the same page as editing the template.
Although this is consistent with how email and text message templates
work, this probably makes it harder to find.

So this commit adds a new page which just lets you rename a template
without changing any other aspects of it.

Also for bilingual letters we will soon have 2 edit pages, one for
English content and one for Welsh content. It would be quite confusing
to either have:
- the edit button on only one of these pages
- the edit button on both (but changing the same field)