Remove things which can’t be changed
We shouldn’t ever have a case where someone can:
- change the ID of a template
- move it to another service
- change an email template into a text message template

I’m not even sure what would happen if you actually tried to make a
request to the API with a change in this data.

Therefore this commit removes those arguments, so the method signature
reflects which fields can be modified.

The API seems happy enough to accept a partial object – the
`post_update_template_schema` in the API doesn’t specify any fields as
`required`:
https://github.com/alphagov/notifications-api/blob/cae3164bfe5bdcdbbbd5117e31f74c3118ef663c/app/template/template_schemas.py#L25-L44