Make chevron in product page breadcrumb svg
Allows us to use the `currentColor` value for the
chevron so it gets set to the same colour as text
in high contrast modes.

This also removes the last use of `separator.png`
and `separator-2x.png` and so deletes both.

Note: the breadcrumb links now get their colour
from the `govuk-link` and `govuk-link--inverse`
classes so don't need any other styles for this.