Return a `Template` model from the `Service` model
At the moment when we have a variable like `template` it’s sometimes a
`dict` straight from the API and it’s sometimes an instance of a model
from utils like `LetterImageTemplate` or `EmailPreviewTemplate`.

This commit makes the model-based approach the default. This is better
because:
- it follows the pattern we use for users, services, branding, etc.
- it means we can add useful methods and properties to the model, which
  are easily reusable across the codebase
- it avoids the looseness of dictionaries where there’s no guarantee
  that specific keys will be present

There’s still more work to do to remove or prevent any interaction with
the underlying `dict`, but this gets us most of the way there.