Remove `LetterPreviewTemplate`
In the old days the admin app could render a letter template in 2 ways:

1. as a series of `<img>` tags, one for each page as the letter
2. as HTML, which could be passed to a HTML-to-PDF convertor for making
   into a PDF

Since then we have moved 2. into the template preview app. So the admin
app only needs to render some `<img>` tags (with the
`LetterImageTemplate` class). Any PDFs or image files that it serves are
rendered by template preview and passed through.

The template preview app still uses the `LetterPreviewTemplate` class to
do this, by rendering HTML which is converted into a PDF. But the admin
app no longer needs to use this class for anything.