Stop error when saving a branding for service with no org
When saving email or letter branding for a service we were also adding
the branding to the pool for the organisation. However, we weren't
checking if a service belonged to an organisation - if it doesn't then
this results in a 404. We now only add the branding to the pool for
services which have an organisation.