Refactor frontend code for last_updated
This code needs to:
- make the baseline of the h1 and last_updated
  line up
- retain the vertical spacing for the h1
- make the h1 and last_updated wrap on smaller
  screens

It's almost impossible to retain the vertical
spacing for the h1 while making its baseline line
up with the last_updated text while they both have
their own vertical margins. I therefore took the
approach of moving those margins onto a parent
container. With flexbox, we can also use the
container to make the baselines of each line up
and to wrap them on smaller screens.