Remove uneeded grid
We were wrapping every template in a grid row and full width column.

This is functionally the same as having no grid at all; the elements
which were inside the grid will still take up the full width.

Removes thiese containing elements makes the code simpler and easier to
follow.