Always put bottom gutter on hint
The only difference between how we display email and text message
templates is whether this little bit of spacing is added.

Before we removed the grid in the previous commit this would have added
more space, because this paragraph was contained in a grid row column,
and the text message ‘sender’ (which has a top margin) was also
contained in a grid row and column. This meant that their margins
added together, giving more space that for other template types,
necessitating the removal of some space.

Now that this paragraph isn’t contained the extra bottom margin will
collapse into the top margin, preserving only the larger bottom margin.

Meaning the spacing will look the same for all template types, which
don’t have the extra top margin introduced by the text message sender.

Discussion on the demerits of top margins best left for another time…