Move padding to individual buttons
With the padding on the row it means that the row will take up space,
even if it contains nothing.

By moving the padding to the individual buttons it will only be shown
if one or both of the buttons are shown.

This means we can simplify the logic and always show the row, and only
have checks for whether the buttons are shown.

This replaces the old, deprecated use of `bottom-gutter` with spacing
modifier classes from GOV.UK Frontend. This is required because
`bottom-gutter` also applies `clear: both`, making the buttons stack
vertically. We don’t want this. The spacing modifier classes from
GOV.UK Frontend do not apply `clear: both`.