Move template archived check
By moving this conditional into `template.html` we can avoid repeatedly
checking it in the template Jinja for each type of Notify template.

In the case that a template is archived all we need to do is render it.
We don’t need to show any of the buttons for editing it, for sending it,
etc.