Combine platform admin search fields
Our current platform admin search page has three boxes on it: one for
services, one for users, and one for UUIDs.

Let's combine these all into a single field so that it's easier to just
drop anything into that field and get the relevant results. We intend to
add organisations to this shortly, but currently the API endpoint for
this doesn't exist.