Make edit letter buttons into secondary buttons
This is so that when we launch the ‘Attach pages’ button all the buttons
on the page will look the same (standard ‘secondary’ buttons from the
GOV.UK Design System).

This is an alternative to making the ‘Attach pages’ button into a blue
link button – having these buttons in a sticky footer would look weird
and isn’t somerthing we’ve done elsewhere.

This means we have two kinds of template page:
- emails and text messages have blue navigation at the top for editing
  and sending
- letter have grey buttons which are located next to the thing they
  modify

While this means the pages are different, it means they are internally
consistent.