Update expressions using '_' to use `do`
Jinja doesn't have syntax to let you just run an
expression that doesn't assign anything by
default. We have occasionally needed to do this,
usually to do things like build lists dynamically,
so had to hack it by assigning the result to the
'_' variable, the "I don't care" variable in
Python.

We added the `do` Jinja extension in this pull
request:

https://github.com/alphagov/notifications-admin/pull/4722/files#diff-9cec7b11237bc29d77a439e81c9b7acfac003d8e8855731eb6bc130b5a8ce602R610

`do` solves the problem of using non-assigning
expressions and makes it explicit what's happening
so we should use that instead. These changes do
that.