Fix comparison between 'thing' and 'name'
We have 2 types of copy-to-clipboard buttons:
1. those that copy a value that is the only one of
   its type on the page (ie. a template ID on a
   template page)
2. those that copy a value which is one of a
   several on the page for that type of thing (ie. a
   reply-to email address on the Reply-to email
   addresses page in service settings)

We use 2 things to differentiate between these
types:
1. the 'thing' (the type of thing it is)
2. the 'name' (the name of that version of the
   thing)

If the 'thing' and the 'name' match, it means
it's the only one of its type on the page so we
leave as is. If they don't, it means we have to
add some text which uses the 'name' to identify
each specific one.

The logic that worked this out didn't compare the
same type of strings so was broken. This fixes it.