Remove fake `image_url` argument values
Previously, `image_url` was a required argument. This meant that we
always needed to provide a value, even if we were never going to
generate any URLs for previewing the template.

Now that’s it’s no longer a required argument, we don’t need to litter
our code with fake values like `example.com`.