Move template type check to top
This check stops you viewing the page to edit a template if your service
doesn’t have that type of template switched on. It doesn’t stop you
actually editing the template if you can craft a post request.

This doesn’t mean anyone can do anything malicious, but the position of
the check in the code doesn’t properly express its intention.

This commit moves it nearer the top, the same as in `add_service`. This
means it will catch both `POST` and `GET` requests.