Fix handling of `400` when trying to send letter
If the API returns a `400` when trying to send an email or text message
(for example because the daily limit has been reached) then we show a
human-friendly version of that error to the user.

This was not working for letters because the view did not know the URL
for showing the PNG preview of the letter. If anyone ended up in this
situation they would see a `500` instead.

This commit adds the PNG preview URL, and a test for this scenario.