Refactor page count to come from template model
It’s very hacky having to do `template.page_count = …` repeatedly.

Instead the `template` instance has all the information it needs to go
and get the page count. It can then return this as a property whenever
it’s needed (and cache it in case of multiple lookups in the same
request).

This is a bit more complicated to implement than might be expected for
two reasons:

1. The existing `Template` model lives in the `notifications-utils`
   package, so we have to subclass it before we can add properties to
   it
2. Page counting works differently for precompiled letters (the page
   count comes from the metadata and needs to be passed in) so we need
   two different subclasses for templated and precompiled letters