Cache the precompiled letter template for a service
The magic template for precompiled letters never changes once it’s
created. So we should never need to make a (relatively) slow call to the
API to get it.

Just in case we ever do change the API response, we can still clear the
template cache manually.

This should make the pages where we view or upload precompiled letters
a tiny bit faster.