Jump to first page of letter attachment
So that you can immediately see the consequence of the action you’ve
just performed.

Relies on the following change from utils version 64.2.0:
* `LetterImageTemplate` now adds a hidden element marking the first page of any attachment

Complete changes: https://github.com/alphagov/notifications-utils/compare/64.1.0...64.2.0

We position the hidden element slightly above the first page of the
attachment. This is so the user sees a little bit of the last page of
the templated letter. This gives some helpful context to understand
where they are.