Be more specific about clear template caches
At the moment, when we edit a template we clear the Redis cache for:
- All templates (because we might have changed the name of a single
  template for example)
- Each individual template belonging to a service

This is overkill, because editing one template shouldn’t mean the cache
needs to be cleared for a different template.

The original intent of doing the pattern delete
(in 01a3df6edce08e3cf0cca42b60dcf95127155497) was to delete the cache
for all _versions_ of an individual template, not every individual
template.

This commit updates the pattern so that it only targets a single
template, but will still delete keys like:
- `…template-1234-version-None`
- `…template-1234-version-1`
- `…template-1234-versions`
- `…template-1234-page-count`

This should improve performance because we will keep most templates
cached for longer.