Make a single checkbox
It doesn’t feel right to have a set of one checkbox. This will change if
we have additional permissions in the future, but for now it’s a simple
yes/no choice.

So this commit removes the fieldset and rewords the label so it makes
sense on its own.