Update create-droplet script
Makes the package and droplet retrieval steps more reliably fetch the
latest package/droplet.

Provides a manifest-droplet.yml file so that we can specify the
buildpack we want to use. Because there are some node-related files in
this repository, without further context cloudfoundry thinks this is a
node app rather than a python app (or at least, it tries to treat it as
a node app before treating it as a python app). We don't want this, so
by using an explicit manifest we can say this is a python-only app.