Defer loading of crests and insignia
Currently this app tries to load the images for government crests and
insignia at startup time.

This means that if the files cannot be found then the app will crash on
startup. This typically happens when the app is restarting because it’s
noticed some changes (for example by checking out a different branch).

The files may be missing because the Gulp task to copy them into the
static folder has not run yet.

Previously this wasn’t a problem because we didn’t delete the old files
before copying the new ones. But as of https://github.com/alphagov/notifications-admin/pull/4710/files
we clean the whole directory before re-copying the assets.

This commit changes the app to not look for the images until it needs
them. By this time the Gulp task should have had a chance to finish
cleaning the folder and copying the images into place.