Auto-prefix page titles with `Error: `
If we pass a form into a template, and it has errors, let's
automatically prefix the page title with `Error: ` as the Design System
suggests (primarily for accessibility).

This automatic processing should cover most cases and be reasonably
enforced by the associated test. But if we need to have multiple forms
on a page, then the developer should override the `pageTitle` block to
handle errors on both forms explicitly.