Fix error summary anchors
When clicking a message in an error summary, it should jump you to the
appropriate form field. These anchors were broken for fields with
multiple inputs, such as radios and checkboxes. GOV.UK Design System
says in these cases, it should take you to the first input:

> For questions that require a user to select one or more options from
> a list using radios or checkboxes, link to the first radio or
> checkbox.

This fixes radio and checkbox fields to provide a custom attribute,
`error_summary_id`, which returns the ID of the first input for the
field. We update the error summary macro to check for this ID and prefer
it, if present.