Allow reordering of form fields
WTForms will iterate over fields on a form in the order that
they are instantiated at runtime (this is generally top-down
as you read through the file). This order is sometimes
different from the order that fields are displayed on the
page. With simple forms this is easy - we can just reorder
the fields in the class declaration. However, we have a
number of forms that have inheritance chains, which makes
reordering more difficult.

Where we are using forms with inheritance and display fields
in a different order than they're created for the form, we
can use `custom_field_order` to rejig the fields. In
particular this may be important to get error message
displaying in error summaries to match the order of forms
visually on the page.