Remove statuspage issues that are now fixed
'our status page ‘subscribe to updates’ link does
not contain programmatically discernible link
text'

The link is now a button (which matches its visual
design) but the same questions apply. The text is
now:
- visible in the page
- available to programs, like assistive tech,
  through accessibility APIs and HTML parsing
- being announced by screen readers, the most
  obvious assistive tech for discerning whether
  what’s in the HTML is usable

'our status page subscribe to updates form fields
do not contain an explicit label'

This was fixed in 2022, as noted by @kr8n3r in:

https://github.com/alphagov/paas-product-pages/pull/143

'our status page subscribe to updates forms do not
indicate to users that they are expandable or
collapsible'

From reading this criterion, I’m confident to
assume that the button saying it’s expanded or
collapsed, which it now does, is enough to pass
this. I also checked that its state is now:
- available to programs, like assistive tech,
  through accessibility APIs and HTML parsing
- being announced by screen readers, the most
  obvious assistive tech for discerning whether
  what’s in the HTML is usable

'our status page subscribe to updates form close
link (x) is not descriptive enough for some users
to determine its function or purpose'

The link is now a button but the same questions
apply. It now has a ‘Close subscribe form’
accessible name which is: The link text is now:
- available to programs, like assistive tech,
  through accessibility APIs and HTML parsing
- being announced by screen readers, the most
  obvious assistive tech for discerning whether
  what’s in the HTML is usable